[PRDP-134] fix node code review and refactor (#48)
* [PRDP-19] feat: defining helm chart tweaks for performance test

* [PRDP-122] Introducing rework to include a node engine

* [PRDP-122] Updated java module of the engine to contact node service, removed legacy implementations, updated node engine to report error codes

* [PRDP-122] feat: Removed handlebar reference from the java function

* [PRDP-122] feat: Updated java module tests and docs, updated code_review.yml to consider the specific java folder for the existing code review step

* [PRDP-122] feat: Updated release-deploy.yml to consider java folder on maven release

* [PRDP-122] feat: Updated release-deploy.yml and code_review.yml

* [PRDP-122] feat: Updated Dockerfile for node and java module

* [PRDP-122] chore: Introducing jest in package.json dependencies, moved util classes to util folder, introduced jest.config.js

* [PRDP-122] chore: Introducing jest in package.json dependencies, moved util classes to util folder, introduced jest.config.js

* [PRDP-122] chore: Updated jest.config.js, Updated package.json

* [PRDP-122] feat: introducing node info endpoint jest testing

* [PRDP-122] feat: Updated node engine tests

* try add sidecar #1

* [PRDP-122] Introducing updated actions with customized implementations

* fix

* fix#2

* fix#3

* fix#4

* fix#5

* fix#6

* [PRDP-122] Updated HttpTriggerGeneratePDFFunction

* fix#7

* fix#8

* fix

* fix##2

* fix##3

* fix##4

* fix##5

* fix##6

* fix##7

* fix##8

* fix##9

* [PRDP-122] Updated Updated release-deploy.yml

* fix

* fix001

* fix001

* fix001

* fix001

* [PRDP-122] feat: refactored app to use express, streamlined structure to use a single Dockerfile

* [PRDP-122] feat: Updated anchore.yml

* [PRDP-122] feat: Updated GeneratePDFServiceImpl

* [PRDP-122] feat: Updated pom.xml

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated Dockerfile

* [PRDP-122] feat: Updated int. tests

* [PRDP-134] feat: Reworked process to manage unzip in java process

* [PRDP-134] feat: Reworked process to manage unzip in java process

* [PRDP-134] feat: Updated pom.xml

* [PRDP-134] feat: Updated ParseRequestBodyServiceImpl

* [PRDP-134] feat: Updated constant.json

* [PRDP-134] feat: Updated code_review.yml

* [PRDP-134] feat: Updated code_review.yml

* [PRDP-134] feat: Updated code_review.yml

* [PRDP-134] feat: Updated code_review.yml

---------
