+++ node/__mocks__/pdf-generate/template.html
@@ -1,0 +1,2 @@
+<html>
+</html>
