+++ node/package.json
@@ -4,19 +4,21 @@
   "description": "",
   "scripts": {
     "start": "node ./pdf-generate/index.js",
+    "dev": "nodemon ./pdf-generate/index.js",
     "test": "jest",
     "test:coverage": "jest --coverage"
   },
   "dependencies": {
     "@anzp/azure-function-multipart": "^3.1.0",
     "applicationinsights": "^1.7.4",
+    "axios": "^1.5.0",
     "express": "^4.18.2",
     "handlebars": "^4.7.7",
     "handlebars-helpers": "^0.10.0",
     "hbs-cli": "^1.4.1",
     "multer": "^1.4.5-lts.1",
-    "puppeteer": "^18.0.2",
-    "unzipper": "^0.10.14"
+    "nodemon": "^3.0.1",
+    "puppeteer": "^18.0.2"
   },
   "devDependencies": {
     "@types/jest": "^24.0.15",
