+++ performance-test/src/test-types/load.json
@@ -13,6 +13,14 @@
     {
       "duration": "1m",
       "target": 20
+    },
+    {
+      "duration": "2m",
+      "target": 20
+    },
+    {
+      "duration": "2m",
+      "target": 0
     }
   ],
   "thresholds": {
