+++ pom.xml
@@ -195,6 +195,12 @@
             <version>4.5.14</version>
         </dependency>
 
+        <dependency>
+            <groupId>net.lingala.zip4j</groupId>
+            <artifactId>zip4j</artifactId>
+            <version>2.11.5</version>
+        </dependency>
+
     </dependencies>
 
     <build>
