+++ src/main/java/it/gov/pagopa/pdf/engine/model/PdfEngineRequest.java
@@ -14,7 +14,9 @@ import java.net.URL;
 @NoArgsConstructor
 public class PdfEngineRequest {
 
-    URL template;
-    String data;
+  String data;
+
+    String workingDirPath;
+
 
 }
