+++ src/test/java/it/gov/pagopa/pdf/engine/client/PdfEngineClientImplTest.java
@@ -42,7 +42,7 @@ class PdfEngineClientImplTest {
         try (InputStream inputStream = FileInputStream.nullInputStream()) {
             template = inputStream.readAllBytes();
 
-            pdfEngineRequest.setTemplate(targetFile.toURI().toURL());
+            pdfEngineRequest.setWorkingDirPath(tempDirectory.getAbsolutePath());
             pdfEngineRequest.setData(new String(template));
         } finally {
             targetFile.deleteOnExit();
@@ -86,8 +86,7 @@ class PdfEngineClientImplTest {
         PdfEngineRequest pdfEngineRequest = new PdfEngineRequest();
         try (InputStream inputStream = FileInputStream.nullInputStream()) {
             template = inputStream.readAllBytes();
-
-            pdfEngineRequest.setTemplate(targetFile.toURI().toURL());
+            pdfEngineRequest.setWorkingDirPath(tempDirectory.getAbsolutePath());
             pdfEngineRequest.setData(new String(template));
         } finally {
             targetFile.deleteOnExit();
@@ -129,7 +128,7 @@ class PdfEngineClientImplTest {
         PdfEngineRequest pdfEngineRequest = new PdfEngineRequest();
         try (InputStream inputStream = FileInputStream.nullInputStream()) {
             template = inputStream.readAllBytes();
-            pdfEngineRequest.setTemplate(targetFile.toURI().toURL());
+            pdfEngineRequest.setWorkingDirPath(tempDirectory.getAbsolutePath());
             pdfEngineRequest.setData(new String(template));
         } finally {
             targetFile.deleteOnExit();
